+++ Netology_IB_Instruments/FeedViewController.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 class FeedViewController: UIViewController {
-    
     let stackView: UIStackView = {
         let stackView = UIStackView()
         stackView.translatesAutoresizingMaskIntoConstraints = false
@@ -16,22 +15,6 @@ class FeedViewController: UIViewController {
         stackView.spacing = 10
         return stackView
     }()
-    func addStackView(){
-        self.stackView.addArrangedSubview(buttonOne)
-        self.stackView.addArrangedSubview(buttonTwo)
-        self.view.addSubview(stackView)
-        NSLayoutConstraint.activate([
-                    self.stackView.centerYAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.centerYAnchor),
-                    
-                    self.stackView.heightAnchor.constraint(equalToConstant: 100),
-                    
-                    self.stackView.leadingAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.leadingAnchor, constant: 50),
-                    
-                    self.stackView.trailingAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.trailingAnchor, constant: -50),
-                    
-                    self.buttonOne.heightAnchor.constraint(equalTo: self.buttonTwo.heightAnchor)
-                ])
-    }
     let buttonOne: UIButton = {
         let button = UIButton()
         button.setTitle("press me", for: .normal)
@@ -44,20 +27,26 @@ class FeedViewController: UIViewController {
         button.backgroundColor = .blue
         return button
     }()
-    func addButton(){
-        self.buttonOne.addTarget(self, action: #selector(gotoPostViewcontroller), for: .touchUpInside)
-        self.buttonTwo.addTarget(self, action: #selector(gotoPostViewcontroller), for: .touchUpInside)
-    }
     override func viewDidLoad() {
         super.viewDidLoad()
-        addButton()
-        addStackView()
+        self.stackView.addArrangedSubview(buttonOne)
+        self.stackView.addArrangedSubview(buttonTwo)
+        self.view.addSubview(stackView)
+        NSLayoutConstraint.activate([
+            self.stackView.centerYAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.centerYAnchor),
+            self.stackView.heightAnchor.constraint(equalToConstant: 100),
+            self.stackView.leadingAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.leadingAnchor, constant: 50),
+            self.stackView.trailingAnchor.constraint(equalTo: self.view.safeAreaLayoutGuide.trailingAnchor, constant: -50),
+            
+            self.buttonOne.heightAnchor.constraint(equalTo: self.buttonTwo.heightAnchor)
+        ])
+        self.buttonOne.addTarget(self, action: #selector(gotoPostViewcontroller), for: .touchUpInside)
+        self.buttonTwo.addTarget(self, action: #selector(gotoPostViewcontroller), for: .touchUpInside)
     }
     @objc func gotoPostViewcontroller(){
         let vc = PostViewController()
         navigationController?.pushViewController(vc, animated: true)
     }
-    
 }
 struct PostTitle {
     var title:String
