+++ Netology_IB_Instruments/InfoViewController.swift
@@ -8,7 +8,6 @@
 import UIKit
 
 class InfoViewController: UIViewController {
-    
     let alertButton: UIButton = {
         let button = UIButton()
         button.translatesAutoresizingMaskIntoConstraints = false
@@ -16,25 +15,17 @@ class InfoViewController: UIViewController {
         button.backgroundColor = .black
         return button
     }()
-    
-    func addAlertButton(){
-        self.view.addSubview(alertButton)
-        alertButton.addTarget(self, action: #selector(tapAlertButton), for: .touchUpInside)
-        NSLayoutConstraint.activate([
-                    alertButton.centerYAnchor.constraint(equalTo: self.view.centerYAnchor),
-                    
-                    alertButton.heightAnchor.constraint(equalToConstant: 50),
-                    
-                    alertButton.leadingAnchor.constraint(equalTo: self.view.leadingAnchor, constant: 50),
-                    
-                    alertButton.trailingAnchor.constraint(equalTo: self.view.trailingAnchor, constant: -50)
-                ])
-    }
-
     override func viewDidLoad() {
         super.viewDidLoad()
         self.view.backgroundColor = .green
-        addAlertButton()
+        self.view.addSubview(alertButton)
+        alertButton.addTarget(self, action: #selector(tapAlertButton), for: .touchUpInside)
+        NSLayoutConstraint.activate([
+            alertButton.centerYAnchor.constraint(equalTo: self.view.centerYAnchor),
+            alertButton.heightAnchor.constraint(equalToConstant: 50),
+            alertButton.leadingAnchor.constraint(equalTo: self.view.leadingAnchor, constant: 50),
+            alertButton.trailingAnchor.constraint(equalTo: self.view.trailingAnchor, constant: -50)
+        ])
     }
     @objc func tapAlertButton(){
         let alertVC = UIAlertController(title: "Error", message: "Fatal Error", preferredStyle: .alert)
@@ -48,5 +39,4 @@ class InfoViewController: UIViewController {
         alertVC.addAction(actionCancel)
         self.present(alertVC, animated: true, completion: nil)
     }
-
 }
