+++ Netology_IB_Instruments/Profile/ProfileHeaderView.swift
@@ -8,29 +8,15 @@
 import UIKit
 
 class ProfileHeaderView: UIView {
-    
     let avatarImageView: UIImageView = {
         let imageView = UIImageView(image: UIImage(named: "elephant.jpg"))
         imageView.layer.borderColor = UIColor.white.cgColor
         imageView.layer.borderWidth = 3
         imageView.layer.cornerRadius = 55
         imageView.layer.masksToBounds = true
+        imageView.translatesAutoresizingMaskIntoConstraints = false
         return imageView
     }()
-    func addAvatarImageView(){
-        self.addSubview(avatarImageView)
-        self.avatarImageView.translatesAutoresizingMaskIntoConstraints = false
-        NSLayoutConstraint.activate([
-            self.avatarImageView.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 16),
-            
-            self.avatarImageView.topAnchor.constraint(equalTo: self.safeAreaLayoutGuide.topAnchor, constant: 16),
-            
-            self.avatarImageView.widthAnchor.constraint(equalToConstant: 110),
-            
-            self.avatarImageView.heightAnchor.constraint(equalToConstant: 110)
-        ])
-    }
-    
     let fullNameLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
@@ -39,20 +25,6 @@ class ProfileHeaderView: UIView {
         label.font = .systemFont(ofSize: 18, weight: .bold)
         return label
     }()
-    
-    func addFullNameLabel(){
-        self.addSubview(fullNameLabel)
-        NSLayoutConstraint.activate([
-            self.fullNameLabel.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 150),
-            
-            self.fullNameLabel.topAnchor.constraint(equalTo: self.safeAreaLayoutGuide.topAnchor, constant: 27),
-            
-            self.fullNameLabel.widthAnchor.constraint(equalToConstant: 150),
-            
-            self.fullNameLabel.heightAnchor.constraint(equalToConstant: 18)
-        ])
-    }
-    
     let statusLabel: UILabel = {
         let label = UILabel()
         label.translatesAutoresizingMaskIntoConstraints = false
@@ -61,20 +33,6 @@ class ProfileHeaderView: UIView {
         label.font = .systemFont(ofSize: 14, weight: .regular)
         return label
     }()
-    
-    func addStatusLabel(){
-        self.addSubview(statusLabel)
-        NSLayoutConstraint.activate([
-            self.statusLabel.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 150),
-            
-            self.statusLabel.topAnchor.constraint(equalTo: self.setStatusButton.topAnchor, constant: -80),
-            
-            self.statusLabel.widthAnchor.constraint(equalToConstant: 190),
-            
-            self.statusLabel.heightAnchor.constraint(equalToConstant: 14)
-        ])
-    }
-    
     let statusTextField: UITextField = {
         let textField = UITextField()
         textField.translatesAutoresizingMaskIntoConstraints = false
@@ -87,21 +45,6 @@ class ProfileHeaderView: UIView {
         textField.textAlignment = .center
         return textField
     }()
-    
-    func addStatusTextField(){
-        self.addSubview(statusTextField)
-        NSLayoutConstraint.activate([
-            self.statusTextField.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 140),
-            
-            self.statusTextField.topAnchor.constraint(equalTo: self.setStatusButton.topAnchor, constant: -50),
-            
-            self.statusTextField.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -16),
-            
-            self.statusTextField.heightAnchor.constraint(equalToConstant: 40)
-        ])
-        self.statusTextField.addTarget(self, action: #selector( statusTextChanged(_:)), for: .editingChanged)
-    }
-    
     let setStatusButton: UIButton = {
         let button = UIButton()
         button.translatesAutoresizingMaskIntoConstraints = false
@@ -115,33 +58,58 @@ class ProfileHeaderView: UIView {
         button.layer.cornerRadius = 4
         return button
     }()
-    func addSetStatusButton(){
-        self.addSubview(setStatusButton)
-        NSLayoutConstraint.activate([
-            self.setStatusButton.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 16),
-            
-            self.setStatusButton.topAnchor.constraint(equalTo: self.avatarImageView.bottomAnchor, constant: 16),
-            
-            self.setStatusButton.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -16),
-            
-            self.setStatusButton.heightAnchor.constraint(equalToConstant: 50)
-        ])
-        self.setStatusButton.addTarget(self, action: #selector(showStatusButtonPressed), for: .touchUpInside)
-    }
-    
     override init(frame: CGRect) {
         super.init(frame: frame)
-        addAvatarImageView()
-        addFullNameLabel()
-        addSetStatusButton()
-        addStatusLabel()
-        addStatusTextField()
+        addSubview()
+        constraints()
+        setupStatusTextField()
+        setupSetStatusButton()
+    }
+    func addSubview(){
+        self.addSubview(avatarImageView)
+        self.addSubview(fullNameLabel)
+        self.addSubview(statusLabel)
+        self.addSubview(statusTextField)
+        self.addSubview(setStatusButton)
+    }
+    func constraints(){
+        NSLayoutConstraint.activate([
+            self.avatarImageView.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 16),
+            self.avatarImageView.topAnchor.constraint(equalTo: self.safeAreaLayoutGuide.topAnchor, constant: 16),
+            self.avatarImageView.widthAnchor.constraint(equalToConstant: 110),
+            self.avatarImageView.heightAnchor.constraint(equalToConstant: 110),
+            
+            self.fullNameLabel.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 150),
+            self.fullNameLabel.topAnchor.constraint(equalTo: self.safeAreaLayoutGuide.topAnchor, constant: 27),
+            self.fullNameLabel.widthAnchor.constraint(equalToConstant: 150),
+            self.fullNameLabel.heightAnchor.constraint(equalToConstant: 18),
+            
+            self.statusLabel.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 150),
+            self.statusLabel.topAnchor.constraint(equalTo: self.setStatusButton.topAnchor, constant: -80),
+            self.statusLabel.widthAnchor.constraint(equalToConstant: 190),
+            self.statusLabel.heightAnchor.constraint(equalToConstant: 14),
+            
+            self.statusTextField.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 140),
+            self.statusTextField.topAnchor.constraint(equalTo: self.setStatusButton.topAnchor, constant: -50),
+            self.statusTextField.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -16),
+            self.statusTextField.heightAnchor.constraint(equalToConstant: 40),
+            
+            self.setStatusButton.leadingAnchor.constraint(equalTo: self.leadingAnchor, constant: 16),
+            self.setStatusButton.topAnchor.constraint(equalTo: self.avatarImageView.bottomAnchor, constant: 16),
+            self.setStatusButton.trailingAnchor.constraint(equalTo: self.trailingAnchor, constant: -16),
+            self.setStatusButton.heightAnchor.constraint(equalToConstant: 50)
+        ])
+    }
+    func setupStatusTextField(){
+        self.statusTextField.addTarget(self, action: #selector( statusTextChanged(_:)), for: .editingChanged)
+    }
+    func setupSetStatusButton(){
+        self.setStatusButton.addTarget(self, action: #selector(showStatusButtonPressed), for: .touchUpInside)
     }
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
     private var statusText:String = ""
-
 @objc
     func showStatusButtonPressed() {
         statusLabel.text = statusText
