+++ Netology_IB_Instruments/Profile/ProfileTableHeaderView.swift
@@ -8,32 +8,23 @@
 import UIKit
 
 class ProfileTableHeaderView: UITableViewHeaderFooterView {
-
     let profileHeaderView: ProfileHeaderView = {
         let view = ProfileHeaderView()
         view.translatesAutoresizingMaskIntoConstraints = false
         view.backgroundColor = .lightGray
         return view
     }()
-    func addProfileHeaderView(){
+    override init(reuseIdentifier: String?) {
+        super.init(reuseIdentifier: reuseIdentifier)
         self.contentView.addSubview(profileHeaderView)
         NSLayoutConstraint.activate([
             self.profileHeaderView.topAnchor.constraint(equalTo: self.contentView.topAnchor),
-            
             self.profileHeaderView.leadingAnchor.constraint(equalTo: self.contentView.leadingAnchor),
-            
             self.profileHeaderView.trailingAnchor.constraint(equalTo: self.contentView.trailingAnchor),
-            
             self.profileHeaderView.heightAnchor.constraint(equalToConstant: 220)
         ])
     }
-    override init(reuseIdentifier: String?) {
-        super.init(reuseIdentifier: reuseIdentifier)
-        addProfileHeaderView()
-    }
-    
     required init?(coder: NSCoder) {
         fatalError("init(coder:) has not been implemented")
     }
-    
 }
