+++ Netology_IB_Instruments/ProfileViewController.swift
@@ -9,39 +9,38 @@
 import UIKit
 
 class ProfileViewController: UIViewController {
-
     let posts:[Post] = [postFirst, postSecond, postThird, postFourth]
     let tableView: UITableView = {
         let view = UITableView(frame: .zero, style: .grouped)
         view.translatesAutoresizingMaskIntoConstraints = false
         return view
     }()
-    func addTableView(){
+    override func loadView() {
+        super.loadView()
         self.view.addSubview(tableView)
-        self.tableView.dataSource = self
-        self.tableView.delegate = self
-        self.tableView.register(PostTableViewCell.self, forCellReuseIdentifier: "idCell")
-        self.tableView.register(ProfileTableHeaderView.self, forHeaderFooterViewReuseIdentifier: "Header")
         NSLayoutConstraint.activate([
             self.tableView.topAnchor.constraint(equalTo: self.view.topAnchor),
-            
             self.tableView.leadingAnchor.constraint(equalTo: self.view.leadingAnchor),
-            
             self.tableView.trailingAnchor.constraint(equalTo: self.view.trailingAnchor),
-            
             self.tableView.bottomAnchor.constraint(equalTo: self.view.bottomAnchor)
         ])
     }
     override func viewDidLoad() {
         super.viewDidLoad()
-        addTableView()
+        setupTableView()
         self.view.backgroundColor = .lightGray
         navigationController?.navigationBar.isHidden = true
     }
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         navigationController?.navigationBar.isHidden = true
-        print("viewWillAppear")
+    }
+    func setupTableView(){
+        self.tableView.dataSource = self
+        self.tableView.delegate = self
+        self.tableView.register(PostTableViewCell.self, forCellReuseIdentifier: "idCell")
+        self.tableView.register(ProfileTableHeaderView.self, forHeaderFooterViewReuseIdentifier: "Header")
+        self.tableView.register(PhotosTableViewCell.self, forCellReuseIdentifier: "idCell2")
     }
 }
 extension ProfileViewController: UITableViewDataSource {
@@ -67,7 +66,7 @@ extension ProfileViewController: UITableViewDataSource {
         case 1:
             let cell = tableView.dequeueReusableCell(withIdentifier: "idCell") as! PostTableViewCell
             cell.authorLabel.text = self.posts[indexPath.row].author
-            cell.postImageView.image = UIImage(named: self.posts[indexPath.row].image)
+            cell.setupCell(image: self.posts[indexPath.row].image)
             cell.postTextView.text = self.posts[indexPath.row].description
             cell.likesLabel.text = "Likes: " + ("\(self.posts[indexPath.row].likes)")
             cell.viewsLabel.text = "Views: " + ("\(self.posts[indexPath.row].views)")
@@ -81,7 +80,6 @@ extension ProfileViewController: UITableViewDataSource {
         return 2
     }
 }
-
 extension ProfileViewController: UITableViewDelegate {
     func tableView(_ tableView: UITableView, estimatedHeightForRowAt indexPath: IndexPath) -> CGFloat {
         return UITableView.automaticDimension
@@ -112,6 +110,6 @@ extension ProfileViewController: UITableViewDelegate {
         default:
             break
         }
-    return nil
+        return nil
     }
 }
